requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
flask-marshmallow==1.2.1
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
marshmallow==3.21.2
mysql-connector-python==8.4.0
packaging==24.0
Werkzeug==3.0.3
zipp==3.18.1
